+++ libs/mobile-components/src/lib/helper/index.ts
@@ -1,2 +1,3 @@
 export * from './storage/';
 export * from './deviceInfo/';
+export * from './mentions'
