+++ libs/mobile-components/src/lib/helper/mentions/index.ts
@@ -1,0 +1,29 @@
+
+export const convertMentionsToText = (text: string) => {
+  const mentionPattern = /{@}\[([^\]]+)\]\(\d+\)|{#}\[([^\]]+)\]\(\d+\)/g;
+  return text.replace(mentionPattern, (match, userMention, hashtagMention) => {
+      if (userMention) {
+          return `@${userMention}`;
+      } else if (hashtagMention) {
+          return `#${hashtagMention}`;
+      } else {
+          return match;
+      }
+  });
+};
+
+export const convertMentionsToData = (text: string) => {
+const mentionPattern = /({@}|{#})\[([^\]]+)\]\((\d+)\)/g;
+const result = [];
+let match;
+while ((match = mentionPattern.exec(text)) !== null) {
+    const prefix = match[1];
+    const mention = match[2];
+    const id = match[3];
+    result.push({
+        id: id,
+        display: `${prefix === '{@}' ? '@' : '#'}${mention}`,
+    });
+}
+return result;
+};
